1.電子柵欄，據點座標修復 2.合約修改初版js 3.合約修改sp 4.短租補傳初版

+++ Web/Scripts/js/ActionJS/StationAndParking/PolygonMaintain.js
@@ -1,7 +1,7 @@
 ﻿
 
 var has = $("#havePolygon").val();
-var myLatLng = { lat: parseInt(lat), lng: parseInt(lng) };
+var myLatLng = { lat: parseFloat(lat), lng: parseFloat(lng) };
 var Color = '#e8e8e8';
 var polyLayersId = [];
 var polyLayers = [];
